+++ src/loginServices.js
@@ -1,5 +1,5 @@
 import localforage from 'localforage';
-import { config, translations } from 'stripes-config';
+import { config, okapi, translations } from 'stripes-config';
 import rtlDetect from 'rtl-detect';
 import moment from 'moment';
 import { loadDayJSLocale } from '@folio/stripes-components';
@@ -126,7 +126,7 @@ export const removeUnauthorizedPathFromSession = () => sessionStorage.removeItem
 export const setUnauthorizedPathToSession = (pathname) => {
   const path = pathname ?? `${window.location.pathname}${window.location.search}`;
   if (!path.startsWith('/logout')) {
-    sessionStorage.setItem(UNAUTHORIZED_PATH, pathname ?? `${window.location.pathname}${window.location.search}`);
+    sessionStorage.setItem(UNAUTHORIZED_PATH, path);
   }
 };
 export const getUnauthorizedPathFromSession = () => sessionStorage.getItem(UNAUTHORIZED_PATH);
@@ -318,8 +318,8 @@ export function getUserLocale(okapiUrl, store, tenant, userId) {
  */
 export function getPlugins(okapiUrl, store, tenant) {
   return fetch(`${okapiUrl}/configurations/entries?query=(module==PLUGINS)`, {
-    headers: getHeaders(tenant, store.getState().okapi.token),
     credentials: 'include',
+    headers: getHeaders(tenant, store.getState().okapi.token),
     mode: 'cors',
   })
     .then((response) => {
@@ -378,15 +378,16 @@ export function getBindings(okapiUrl, store, tenant) {
 /**
  * loadResources
  * return a promise that retrieves the tenant's locale, user's locale,
- * plugins, and key-bindings from mod-configuration.
- * @param {string} okapiUrl
+ * plugins, and key-bindings from mod-configuration, as well as retreiving
+ * module information directly from okapi.
+ *
  * @param {redux store} store
  * @param {string} tenant
  * @param {string} userId user's UUID
  *
  * @returns {Promise}
  */
-function loadResources(okapiUrl, store, tenant, userId) {
+function loadResources(store, tenant, userId) {
   let promises = [];
 
   // tenant's locale, plugin, bindings, and user's locale are all stored
@@ -394,10 +395,10 @@ function loadResources(okapiUrl, store, tenant, userId) {
   // read-permission for configuration entries.
   if (canReadConfig(store)) {
     promises = [
-      getLocale(okapiUrl, store, tenant),
-      getUserLocale(okapiUrl, store, tenant, userId),
-      getPlugins(okapiUrl, store, tenant),
-      getBindings(okapiUrl, store, tenant),
+      getLocale(okapi.url, store, tenant),
+      getUserLocale(okapi.url, store, tenant, userId),
+      getPlugins(okapi.url, store, tenant),
+      getBindings(okapi.url, store, tenant),
     ];
   }
 
@@ -428,16 +429,31 @@ export function spreadUserWithPerms(userWithPerms) {
     ...userWithPerms?.user?.personal,
   };
 
-  // remap userWithPerms.permissions.permissions from an array shaped like
-  //   [{ "permissionName": "foo", ... }]
-  // to an object shaped like
-  //   { foo: true, ...}
-  const perms = {};
+  // remap data's array of permission-names to set with
+  // permission-names for keys and `true` for values.
+  //
+  // userWithPerms is shaped differently depending on the API call
+  // that generated it.
+  // in community-folio, /login sends data like [{ "permissionName": "foo" }]
+  //   and includes both directly and indirectly assigned permissions
+  // in community-folio, /_self sends data like ["foo", "bar", "bat"]
+  //   but only includes directly assigned permissions
+  // in community-folio, /_self?expandPermissions=true sends data like [{ "permissionName": "foo" }]
+  //   and includes both directly and indirectly assigned permissions
+  // in eureka-folio, /_self sends data like ["foo", "bar", "bat"]
+  //   and includes both directly and indirectly assigned permissions
+  //
+  // we'll parse it differently depending on what it looks like.
+  let perms = {};
   const list = userWithPerms?.permissions?.permissions;
   if (list && Array.isArray(list) && list.length > 0) {
-    list.forEach(p => {
-      perms[p.permissionName] = true;
-    });
+    // shaped like this ["foo", "bar", "bat"] or
+    // shaped like that [{ "permissionName": "foo" }]?
+    if (typeof list[0] === 'string') {
+      perms = Object.assign({}, ...list.map(p => ({ [p]: true })));
+    } else {
+      perms = Object.assign({}, ...list.map(p => ({ [p.permissionName]: true })));
+    }
   }
 
   return { user, perms };
@@ -483,10 +499,12 @@ export async function logout(okapiUrl, store, queryClient) {
     fetch(`${okapiUrl}/authn/logout`, {
       method: 'POST',
       mode: 'cors',
-      credentials: 'include'
+      credentials: 'include',
+      headers: getHeaders(store.getState()?.okapi?.tenant),
     })
     :
     Promise.resolve();
+
   return logoutPromise
     // clear private-storage
     .then(() => {
@@ -535,15 +553,14 @@ export async function logout(okapiUrl, store, queryClient) {
  * Dispatch the session object, then return a Promise that fetches
  * and dispatches tenant resources.
  *
- * @param {string} okapiUrl
- * @param {object} store
- * @param {string} tenant
- * @param {string} token
+ * @param {object} store redux store
+ * @param {string} tenant tenant name
+ * @param {string} token access token [deprecated; prefer folioAccessToken cookie]
  * @param {*} data
  *
  * @returns {Promise}
  */
-export function createOkapiSession(okapiUrl, store, tenant, token, data) {
+export function createOkapiSession(store, tenant, token, data) {
   // clear any auth-n errors
   store.dispatch(setAuthError(null));
 
@@ -555,7 +572,7 @@ export function createOkapiSession(okapiUrl, store, tenant, token, data) {
 
   store.dispatch(setCurrentPerms(perms));
 
-  // if we can't parse tokenExpiration data, e.g. because data comes from `/bl-users/_self`
+  // if we can't parse tokenExpiration data, e.g. because data comes from `.../_self`
   // which doesn't provide it, then set an invalid AT value and a near-future (+10 minutes) RT value.
   // the invalid AT will prompt an RTR cycle which will either give us new AT/RT values
   // (if the RT was valid) or throw an RTR_ERROR (if the RT was not valid).
@@ -581,11 +598,38 @@ export function createOkapiSession(okapiUrl, store, tenant, token, data) {
   localStorage.setItem(SESSION_NAME, 'true');
 
   return localforage.setItem('loginResponse', data)
-    .then(() => localforage.setItem(SESSION_NAME, okapiSess))
+    .then(localforage.getItem(SESSION_NAME))
+    .then(sessionData => {
+      // for keycloak-based logins, token-expiration data was already
+      // pushed to storage, so we pull it out and reuse it here.
+      // for legacy logins, token-expiration data is available in the
+      // login response.
+      if (sessionData?.tokenExpiration) {
+        okapiSess.tokenExpiration = sessionData.tokenExpiration;
+      } else if (data.tokenExpiration) {
+        okapiSess.tokenExpiration = {
+          atExpires: new Date(data.tokenExpiration.accessTokenExpiration).getTime(),
+          rtExpires: new Date(data.tokenExpiration.refreshTokenExpiration).getTime(),
+        };
+      } else {
+        // somehow, we ended up here without any legit token-expiration values.
+        // that's not great, but in theory we only ended up here as a result
+        // of logging in, so let's punt and assume our cookies are valid.
+        // set an expired AT and RT 10 minutes into the future; the expired
+        // AT will kick off RTR, and on success we'll store real values as a result,
+        // or on failure we'll get kicked out. no harm, no foul.
+        okapiSess.tokenExpiration = {
+          atExpires: -1,
+          rtExpires: Date.now() + (10 * 60 * 1000),
+        };
+      }
+
+      return localforage.setItem(SESSION_NAME, okapiSess);
+    })
     .then(() => {
       store.dispatch(setIsAuthenticated(true));
       store.dispatch(setSessionData(okapiSess));
-      return loadResources(okapiUrl, store, sessionTenant, user.id);
+      return loadResources(store, sessionTenant, user.id);
     });
 }
 
@@ -674,22 +718,33 @@ export function handleLoginError(dispatch, resp) {
 /**
  * processOkapiSession
  * create a new okapi session with the response from either a username/password
- * authentication request or a bl-users/_self request.
- * @param {string} okapiUrl
+ * authentication request or a .../_self request.
+ * response body is shaped like
+ * {
+    'access_token': 'SOME_STRING',
+    'expires_in': 420,
+    'refresh_expires_in': 1800,
+    'refresh_token': 'SOME_STRING',
+    'token_type': 'Bearer',
+    'not-before-policy': 0,
+    'session_state': 'SOME_UUID',
+    'scope': 'profile email'
+ * }
+ *
  * @param {redux store} store
  * @param {string} tenant
  * @param {Response} resp HTTP response
  *
  * @returns {Promise} resolving with login response body, rejecting with, ummmmm
  */
-export function processOkapiSession(okapiUrl, store, tenant, resp, ssoToken) {
-  const token = resp.headers.get('X-Okapi-Token') || ssoToken;
+export function processOkapiSession(store, tenant, resp, ssoToken) {
   const { dispatch } = store;
 
   if (resp.ok) {
     return resp.json()
       .then(json => {
-        return createOkapiSession(okapiUrl, store, tenant, token, json)
+        const token = resp.headers.get('X-Okapi-Token') || json.access_token || ssoToken;
+        return createOkapiSession(store, tenant, token, json)
           .then(() => json);
       })
       .then((json) => {
@@ -703,7 +758,7 @@ export function processOkapiSession(okapiUrl, store, tenant, resp, ssoToken) {
 
 /**
  * validateUser
- * return a promise that fetches from bl-users/_self.
+ * return a promise that fetches from .../_self.
  * if successful, dispatch the result to create a session
  * if not, clear the session and token.
  *
@@ -716,7 +771,8 @@ export function processOkapiSession(okapiUrl, store, tenant, resp, ssoToken) {
  */
 export function validateUser(okapiUrl, store, tenant, session) {
   const { token, tenant: sessionTenant = tenant } = session;
-  return fetch(`${okapiUrl}/bl-users/_self?expandPermissions=true`, {
+  const usersPath = okapi.authnUrl ? 'users-keycloak' : 'bl-users';
+  return fetch(`${okapiUrl}/${usersPath}/_self?expandPermissions=true`, {
     headers: getHeaders(sessionTenant, token),
     credentials: 'include',
     mode: 'cors',
@@ -748,7 +804,7 @@ export function validateUser(okapiUrl, store, tenant, session) {
           tokenExpiration: session.tokenExpiration
         }));
 
-        return loadResources(okapiUrl, store, sessionTenant, user.id);
+        return loadResources(store, sessionTenant, user.id);
       });
     } else {
       store.dispatch(clearCurrentUser());
@@ -781,7 +837,10 @@ export function checkOkapiSession(okapiUrl, store, tenant) {
       return sess?.user?.id ? validateUser(okapiUrl, store, tenant, sess) : null;
     })
     .then(() => {
-      return getSSOEnabled(okapiUrl, store, tenant);
+      if (store.getState().discovery?.interfaces?.['login-saml']) {
+        return getSSOEnabled(okapiUrl, store, tenant);
+      }
+      return Promise.resolve();
     })
     .finally(() => {
       store.dispatch(setOkapiReady());
@@ -792,7 +851,7 @@ export function checkOkapiSession(okapiUrl, store, tenant) {
  * requestLogin
  * authenticate with a username and password. return a promise that posts the values
  * and then processes the result to begin a session.
- * @param {string} okapiUrl
+ * @param {object} okapi object from stripes.config.js
  * @param {redux store} store
  * @param {string} tenant
  * @param {object} data
@@ -800,15 +859,30 @@ export function checkOkapiSession(okapiUrl, store, tenant) {
  * @returns {Promise}
  */
 export function requestLogin(okapiUrl, store, tenant, data) {
-  const loginPath = config.useSecureTokens ? 'login-with-expiry' : 'login';
-  return fetch(`${okapiUrl}/bl-users/${loginPath}?expandPermissions=true&fullPermissions=true`, {
-    body: JSON.stringify(data),
-    credentials: 'include',
-    headers: { 'X-Okapi-Tenant': tenant, 'Content-Type': 'application/json' },
-    method: 'POST',
-    mode: 'cors',
-  })
-    .then(resp => processOkapiSession(okapiUrl, store, tenant, resp));
+  // got Keycloak?
+  if (okapi.authnUrl) {
+    return fetch(okapi.authnUrl, {
+      method: 'POST',
+      body: new URLSearchParams({
+        ...data,
+        'grant_type': 'password',
+        'client_id': okapi.clientId,
+        'client_secret': okapi.clientSecret,
+      })
+    })
+      .then(resp => processOkapiSession(store, tenant, resp));
+  } else {
+    // legacy built-in authentication
+    const loginPath = config.useSecureTokens ? 'login-with-expiry' : 'login';
+    return fetch(`${okapiUrl}/bl-users/${loginPath}?expandPermissions=true&fullPermissions=true`, {
+      body: JSON.stringify(data),
+      credentials: 'include',
+      headers: { 'X-Okapi-Tenant': tenant, 'Content-Type': 'application/json' },
+      method: 'POST',
+      mode: 'cors',
+    })
+      .then(resp => processOkapiSession(store, tenant, resp));
+  }
 }
 
 /**
@@ -822,8 +896,9 @@ export function requestLogin(okapiUrl, store, tenant, data) {
  * @returns {Promise} Promise resolving to the response of the request
  */
 function fetchUserWithPerms(okapiUrl, tenant, token, rtrIgnore = false) {
+  const usersPath = okapi.authnUrl ? 'users-keycloak' : 'bl-users';
   return fetch(
-    `${okapiUrl}/bl-users/_self?expandPermissions=true&fullPermissions=true`,
+    `${okapiUrl}/${usersPath}/_self?expandPermissions=true&fullPermissions=true`,
     {
       headers: getHeaders(tenant, token),
       rtrIgnore,
@@ -842,7 +917,15 @@ function fetchUserWithPerms(okapiUrl, tenant, token, rtrIgnore = false) {
  */
 export function requestUserWithPerms(okapiUrl, store, tenant, token) {
   return fetchUserWithPerms(okapiUrl, tenant, token, !token)
-    .then(resp => processOkapiSession(okapiUrl, store, tenant, resp, token));
+    .then((resp) => {
+      if (resp.ok) {
+        return processOkapiSession(store, tenant, resp, token);
+      } else {
+        return resp.json().then((error) => {
+          throw error;
+        });
+      }
+    });
 }
 
 /**
@@ -892,14 +975,14 @@ export function updateUser(store, data) {
  * updateTenant
  * 1. prepare user info for requested tenant
  * 2. update okapi session
- * @param {object} okapi
+ * @param {object} okapiConfig
  * @param {string} tenant
  *
  * @returns {Promise}
  */
-export async function updateTenant(okapi, tenant) {
+export async function updateTenant(okapiConfig, tenant) {
   const okapiSess = await getOkapiSession();
-  const userWithPermsResponse = await fetchUserWithPerms(okapi.url, tenant, okapi.token);
+  const userWithPermsResponse = await fetchUserWithPerms(okapiConfig.url, tenant, okapiConfig.token);
   const userWithPerms = await userWithPermsResponse.json();
   await localforage.setItem(SESSION_NAME, { ...okapiSess, tenant, ...spreadUserWithPerms(userWithPerms) });
 }