+++ test/jest/__mock__/stripesComponents.mock.js
@@ -70,7 +70,13 @@ jest.mock('@folio/stripes-components', () => ({
   Label: jest.fn(({ children, ...rest }) => (
     <span {...rest}>{children}</span>
   )),
+  List: jest.fn(({ items, itemFormatter = (item, i) => (<li key={i}>{item}</li>) }) => (
+    <ul>
+      [{items?.map((item, i) => itemFormatter(item, i))}]
+    </ul>
+  )),
   Loading: () => <div>Loading</div>,
+  LoadingView: () => <div>LoadingView</div>,
   MessageBanner: jest.fn(({ show, children }) => { return show ? <>{children}</> : <></>; }),
 
   // oy, dismissible. we need to pull it out of props so it doesn't